Refactor: transaction 파일들 폴더 이름, 위치 변경
