+++ src/main/java/service/transactionList/TransactionListService.java
@@ -1,0 +1,20 @@
+package service.transactionList;
+
+import domain.TransactionVO;
+
+import java.util.List;
+
+public interface TransactionListService {
+
+    // 유저 ID로 모든 거래 내역 조회
+    List<TransactionVO> getAllTransactionsByUserId(int userId);
+
+    // 정렬 옵션으로 거래 내역 조회 (amount, memo, timestamp 등, asc/desc)
+    List<TransactionVO> getTransactionsByUserIdSorted(int userId, String sortBy, boolean asc);
+
+
+    // 거래 내역 N건 조회 (최신 N건)
+    List<TransactionVO> getRecentTransactionsByUserId(int userId, int limit);
+
+
+}
